code update

+++ app/src/main/java/in/dreamcode/roomdemo/Dao/DbDao.kt
@@ -18,6 +18,6 @@ interface DbDao {
     fun getAllNote(): List<NoteModel>
 
     @Query("SELECT * FROM NoteModel WHERE id= :noteId")
-    fun getNoteById(noteId: Int): NoteModel
+    fun getNoteById(noteId: Long): NoteModel
 
 }
